+++ R/crunch-page.R
@@ -5,7 +5,9 @@
 #' Crunch UI style and keep your users authenticated. When building shiny apps
 #' with Crunch datasets, use these instead of [shiny::fluidPage()],
 #' [shiny::fillPage()] or [shiny::navbarPage()].
-#' @param title the title to be displayed in the navigation bar. Defaults to NULL.
+#'
+#' These are no longer necessary. Just use the `shiny` ones and it just works.
+#' These functions are left here for backwards compatibility.
 #' @param ... arguments passed to `fluidPage`, `fillPage` or `navbarPage`
 #' @return The result of `fluidPage`, `fillPage` or `navbarPage`
 #' @export
@@ -33,13 +35,7 @@
 #'     )
 #' )
 #' }
-crunchPage <- function (...) {
-    fluidPage(
-        loadCrunchAssets(),
-        crunchAuthPlaceholder(),
-        ...
-    )
-}
+crunchPage <- function (...) fluidPage(...)
 
 #' @rdname crunchPage
 #' @export
@@ -47,37 +43,55 @@ crunchFluidPage <- crunchPage
 
 #' @rdname crunchPage
 #' @export
-crunchFillPage <- function (...) {
-    fillPage(
-        loadCrunchAssets(),
-        crunchAuthPlaceholder(),
-        ...
-    )
-}
+crunchFillPage <- function (...) fillPage(...)
 
 #' @rdname crunchPage
 #' @export
-crunchNavbarPage <- function(title = NULL, ...) {
-  navbarPage(title = title,
-    loadCrunchAssets(),
-    crunchAuthPlaceholder(),
-    ...
-  )
-}
+crunchNavbarPage <- function (...) navbarPage(...)
 
-#' @importFrom shiny includeCSS includeScript tags
-loadCrunchAssets <- function () {
-    tags$head(
-        tags$link(rel="stylesheet", type="text/css",
-            href="https://app.crunch.io/styles.css"),
-        includeCSS(system.file("extra.css", package="crunchy")),
-        includeScript(system.file("extra.js", package="crunchy"))
+#' @importFrom shiny div includeCSS includeScript tags
+injectCrunchAssets <- function () {
+    suppressMessages(
+        trace(
+            "bootstrapPage",
+            where=shiny::fillPage,
+            tracer=crunchAssets,
+            print=FALSE
+        )
     )
 }
 
-#' @importFrom shiny div
-crunchAuthPlaceholder <- function () {
-    div(class = "form-group shiny-input-container",
-        style = "display: none;",
-        tags$input(id = "token", type = "text", class = "form-control", value = ""))
-}
+crunchAssets <- quote({
+    attachDependencies <- function (x, ...) {
+        # Prepend a bit of HTML before whatever the user supplied but after
+        # the global things--as if these were the first things in the user's
+        # list
+        x[[length(x)]] <- c(
+            list(
+                # Load Crunch assets
+                tags$head(
+                    tags$link(
+                        rel="stylesheet",
+                        type="text/css",
+                        href="https://app.crunch.io/styles.css"
+                    ),
+                    includeCSS(system.file("extra.css", package="crunchy")),
+                    includeScript(system.file("extra.js", package="crunchy"))
+                ),
+                # Add a placeholder for Crunch auth
+                div(
+                    class = "form-group shiny-input-container",
+                    style = "display: none;",
+                    tags$input(
+                        id = "token",
+                        type = "text",
+                        class = "form-control",
+                        value = ""
+                    )
+                )
+            ),
+            x[[length(x)]]
+        )
+        htmltools::attachDependencies(x, ...)
+    }
+})