+++ R/crunchy.R
@@ -1,0 +1,4 @@
+.onLoad <- function (lib, pkgname="crunchy") {
+    injectCrunchAssets()
+    invisible()
+}